[step4] 로또(수동) 리팩토링

+++ src/main/kotlin/lotto/step4/domain/RankClassifier.kt
@@ -5,11 +5,11 @@ object RankClassifier {
         lottos: List<Lotto>,
         winningNumbers: Set<LottoNumber>,
         bonusNumber: LottoNumber,
-    ): Map<Rank, Int> {
+    ): Map<Rank, Long> {
         return lottos.groupBy { lotto ->
             val matchCount = lotto.numbers.intersect(winningNumbers).size
             val hasBonusNumber = lotto.numbers.contains(bonusNumber) && matchCount == 5 // 보너스 번호는 2등 조건에만 적용
             Rank.valueOf(matchCount, hasBonusNumber)
-        }.mapValues { it.value.size }
+        }.mapValues { it.value.size.toLong() }
     }
 }
