+++ src/main/kotlin/lotto/step4/domain/WinningStatistics.kt
@@ -1,6 +1,6 @@
 package lotto.step4.domain
 
 data class WinningStatistics(
-    val rankMap: Map<Rank, Int>,
+    val rankMap: Map<Rank, Long>,
     val profit: Double,
 )
